More circular dependencies removed
