+++ src/ducks/Studio/Template/index.js
@@ -22,7 +22,7 @@ import DialogFormRenameTemplate from './Dialog/RenameTemplate'
 import DialogLoadTemplate from './Dialog/LoadTemplate'
 import DeleteTemplate from './Dialog/Delete/DeleteTemplate'
 import ShareTemplate from './Share/ShareTemplate'
-import { isUserAuthorOfTemplate } from './Dialog/LoadTemplate/Template'
+import { isUserAuthorOfTemplate } from './Dialog/LoadTemplate/utils'
 import { useKeyboardCmdShortcut } from '../hooks'
 import { parseSharedWidgets, translateMultiChartToWidgets } from '../url/parse'
 import { normalizeWidgets } from '../url/generate'
