+++ src/ducks/Watchlists/Widgets/Filter/formatters.js
@@ -1,0 +1,2 @@
+export const percentServerValueFormatter = value => value / 100
+export const percentValueFormatter = value => value * 100
