+++ src/index.js
@@ -10,9 +10,7 @@ import { ApolloProvider } from 'react-apollo'
 import { ConnectedRouter } from 'react-router-redux'
 import App from './App'
 import { client } from './apollo'
-import { store } from './redux'
-import reducers from './reducers/rootReducers.js'
-import epics from './epics/rootEpics.js'
+import { store, history } from './redux'
 import { saveState } from './utils/localStorage'
 import { getAPIUrl, isNotSafari } from './utils/utils'
 import detectNetwork from './utils/detectNetwork'
