+++ src/pages/feed/GeneralFeed/PulseInsights/PulseInsights.js
@@ -6,7 +6,7 @@ import Loader from '@santiment-network/ui/Loader/Loader'
 import { PULSE_INSIGHTS_BY_PAGE_QUERY } from '../../../../queries/InsightsGQL'
 import PulseInsightWrapper from '../../../../components/Insight/PulseInsight'
 import { isBottom } from '../utils'
-import { EmptyFeed } from '../GeneralFeed'
+import EmptyFeed from '../EmptyFeed'
 import { groupByDates, RenderFeedGroups } from '../FeedList/FeedList'
 import styles from '../FeedItemRenderer/FeedItemRenderer.module.scss'
 import feedlistStyles from '../FeedList/FeedList.module.scss'
