+++ src/pages/feed/GeneralFeed/locations.js
@@ -1,0 +1,3 @@
+export const baseLocation = '/feed'
+export const pulseLocation = `${baseLocation}/pulse`
+export const personalLocation = `${baseLocation}/personal`
